Done some rework on the UI. Lot's left to do.

+++ Swipper2/src/com/globant/labs/swipper2/utils/DroidUtils.java
@@ -4,7 +4,7 @@ import android.content.Context;
 
 public class DroidUtils {
 
-	public static int dpToPx(int dp, Context context) {
+	public static int dpToPx(float dp, Context context) {
 		float scale = context.getResources().getDisplayMetrics().density;
 		int dpAsPixels = (int) (dp * scale + 0.5f);
 		return dpAsPixels;
